==> jupyter notebook/preprocessing/2. providers.ipynb <==
# Unique Providers

from pathlib import Path
import os

os.chdir(Path(os.getcwd()).parent)

from os.path import join

import pandas as pd

## Load news data

root_path = os.getcwd()
data_path = join(os.getcwd(), "data")
raw_data_path = join(data_path, "raw")
pre_data_path = join(data_path, "preprocessing")

news_df = pd.read_csv(join(raw_data_path, "news_data_20000.csv"))

providers = news_df["provider"].unique()

providers

len(providers)

## Load Naver Providers

naver_df = pd.read_csv(join(raw_data_path, "providers.csv"))

naver_providers = naver_df["이름"].unique()

## 네이버 / 빅카인즈 언론사 비교

for provider in providers:
    if provider not in naver_providers:
        print(provider)

## 빅카인즈 언론사 저장

provider_df = pd.DataFrame(columns=["이름", "분류", "지역"])

for provider in providers:
    if provider in naver_providers:
        naver_row = naver_df[naver_df["이름"] == provider]
        provider_df = pd.concat([provider_df, naver_row], ignore_index=True)

울산매일, 한라일보 추가 저장

add1 = {"이름": "울산매일", "분류": "지역지", "지역": "울산"}
add2 = {"이름": "한라일보", "분류": "지역지", "지역": "제주"}

provider_df = pd.concat([provider_df, pd.DataFrame([add1, add2])], ignore_index=True)

provider_df.head()

provider_df.shape

provider_df.to_csv(join(pre_data_path, "providers.csv"), index=False, encoding="utf-8-sig")
