==> jupyter notebook/preprocessing/3. categories.ipynb <==
# 뉴스 카테고리

from pathlib import Path
import os

os.chdir(Path(os.getcwd()).parent)

from os.path import join
from collections import defaultdict
import ast

import pandas as pd

## Load news data

root_path = os.getcwd()
data_path = join(os.getcwd(), "data")

news_df = pd.read_csv(join(data_path, "news_data_20000.csv"))

categories = news_df["category"].unique()
categories[:5]

category_set = set()

for category in categories:
    category_list = ast.literal_eval(category)
    for elem in category_list:
        category_set.add(elem)

categories = list(category_set)

len(categories)

categories[:5]

## 대/중/소 카테고리 분류

majors = set()

for category in categories:
    elems = category.split(">")
    majors.add(elems[0])

majors = list(majors)

len(majors)

majors

category_dict = defaultdict(set)

for category in categories:
    elems = category.split(">")
    category_dict[elems[0]].add(elems[1])

for major_category, minors in category_dict.items():
    print(major_category)
    print(minors)
    print()

## 대분류/중분류 저장

category_df = pd.DataFrame(columns=["대분류", "중분류"])

for major_category, minors in category_dict.items():
    for minor in list(minors):
        category_df = pd.concat([category_df, pd.DataFrame([{"대분류": major_category, "중분류": minor}])], ignore_index=True)

category_df.head()

category_df.to_csv(join(data_path, "categories.csv"), index=False, encoding="utf-8-sig")
