==> jupyter notebook/preprocessing/4-2. bylines - korean.ipynb <==
# 기자 전처리1

from pathlib import Path
import os

os.chdir(Path(os.getcwd()).parent)

from os.path import join

import pandas as pd

## Load byline data

root_path = os.getcwd()
data_path = join(os.getcwd(), "data", "preprocessing")
byline_path = join(data_path, "bylines")

byline_df = pd.read_csv(join(byline_path, "bylines.csv"))
byline_df.head()

## 이름 분류
- normal: 이름이 3글자인 경우
- abnormal: 그외

names = byline_df["이름"]

normal_names, abnormal_names = set(), set()

for name in names:
    if len(name) == 3:
        normal_names.add(name)
    else:
        abnormal_names.add(name)

len(normal_names), len(abnormal_names)

normal_names = sorted(normal_names)
abnormal_names = sorted(abnormal_names)

name_df = pd.DataFrame(columns=["이름"])

for name in normal_names:
    name_df = pd.concat([name_df, pd.DataFrame([{"이름": name}])], ignore_index=True)

name_df.to_csv(join(byline_path, "names.csv"), index=False, encoding="utf-8-sig")

name_to_process_df = pd.DataFrame(columns=["이름"])

for name in abnormal_names:
    name_to_process_df = pd.concat([name_to_process_df, pd.DataFrame([{"이름": name}])], ignore_index=True)

name_to_process_df.to_csv(join(byline_path, "abnormal_names.csv"), index=False, encoding="utf-8-sig")

## 전처리 할 이름 분류

name_to_process_df.head()

### Case1 - 개행이 있는 경우

names = abnormal_names
names_with_newline, others = [], []

for name in names:
    if len(name.splitlines()) > 1:
        names_with_newline.append(name)
    else:
        others.append(name)

len(names_with_newline), len(others)

names_newline_df = pd.DataFrame()
names_newline_df["이름"] = names_with_newline
names_newline_df.to_csv(join(byline_path, "names_newline.csv"), index=False, encoding="utf-8-sig")

others_df = pd.DataFrame()
others_df["이름"] = others
others_df.to_csv(join(byline_path, "names_others.csv"), index=False, encoding="utf-8-sig")
