==> jupyter notebook/preprocessing/4-3. bylines.ipynb <==
# 기자 전처리 2

from pathlib import Path
import os

os.chdir(Path(os.getcwd()).parent)

from os.path import join

import pandas as pd
import numpy as np

## Load Others data

root_path = os.getcwd()
data_path = join(os.getcwd(), "data", "preprocessing")
byline_path = join(data_path, "bylines")

byline_df = pd.read_csv(join(byline_path, "names_others.csv"))
byline_df.head()

### Case 2
ooo 기자

byline_df["정규화"] = np.nan

def iter_loop(name: str) -> tuple[bool, str]:
    name_splits = name.split(" ")
    
    if len(name_splits) == 2:
        if name_splits[1] == "기자":
            return True, name_splits[0]
        else:
            return False, name
    else:
        return False, name

others = []

for index, row in byline_df.iterrows():
    is_reporter, name = iter_loop(row["이름"])

    if is_reporter:
        byline_df.at[index, "정규화"] = name
    else:
        others.append(name)

len(others)

byline_df.to_csv(join(byline_path, "names_others.csv"), index=False, encoding="utf-8-sig")

others_df = pd.DataFrame()
others_df["이름"] = others
others_df.to_csv(join(byline_path, "names_others1.csv"), index=False, encoding="utf-8-sig")
