==> jupyter notebook/preprocessing/4-4. bylines.ipynb <==
# 기자 전처리 3

### csv 정보
- bylines.csv: 원본 기자명
- names.csv: 3글자 기자명
  - 수동 정규화 - 단체 이름 제거
- names_newline.csv: 개행 문자 포함한 기자명
  - 수동 정규화
- names_others.csv
  - `OOO 기자` 정규화
  - `OOO` 수동 정규화
- names_others1.csv
  - 나머지 이름 수동 정규화

from pathlib import Path
import os

os.chdir(Path(os.getcwd()).parent)

from os.path import join

import pandas as pd
import numpy as np

root_path = os.getcwd()
data_path = join(os.getcwd(), "data", "preprocessing")
byline_path = join(data_path, "bylines")
reporter_path = join(data_path, "bylines.csv")

## 1. bylines -> reporters
- 수동 복사

reporter_df = pd.read_csv(reporter_path)

reporter_df.head()

reporter_df.shape

reporter_df["정규화"] = np.nan

reporter_df= reporter_df.astype({"정규화": "object"})
reporter_df.dtypes

## 2. names → reporters

names_df = pd.read_csv(join(byline_path, "names.csv"))

names_df.head()

for index, row in names_df.iterrows():
    name, new_name = row["이름"], row["정규화"]
    if new_name is not np.nan:
        reporter_df.loc[reporter_df["이름"] == name, "정규화"] = new_name

## 3. names_newline → reporters

names_newline_df = pd.read_csv(join(byline_path, "names_newline.csv"))

names_newline_df.head()

for index, row in names_newline_df.iterrows():
    name, new_name = row["이름"], row["정규화"]
    if new_name is not np.nan:
        reporter_df.loc[reporter_df["이름"] == name, "정규화"] = new_name

## 4. names_others → reporters

others_df = pd.read_csv(join(byline_path, "names_others.csv"))

others_df.head()

for index, row in others_df.iterrows():
    name, new_name = row["이름"], row["정규화"]
    if new_name is not np.nan:
        reporter_df.loc[reporter_df["이름"] == name, "정규화"] = new_name

## 5. names_others1 → reporters

others1_df = pd.read_csv(join(byline_path, "names_others1.csv"))

others1_df.head()

for index, row in others1_df.iterrows():
    name, new_name = row["이름"], row["정규화"]
    if new_name is not np.nan:
        reporter_df.loc[reporter_df["이름"] == name, "정규화"] = new_name

## 6. 저장

reporter_df.shape

reporter_df.head()

reporter_df.to_csv(reporter_path, index=False, encoding="utf-8-sig")
