==> jupyter notebook/preprocessing/5. reporters.ipynb <==
# Reporters

from pathlib import Path
import os

os.chdir(Path(os.getcwd()).parent)

from os.path import join

import pandas as pd
import numpy as np

root_path = os.getcwd()
data_path = join(os.getcwd(), "data", "preprocessing")
byline_path = join(data_path, "bylines.csv")
reporter_path = join(data_path, "reporters.csv")

byline_df = pd.read_csv(byline_path)
raw_names = sorted(byline_df["정규화"].dropna().unique())
len(raw_names)

names = set()

for raw_name in raw_names:
    for name in raw_name.split("||"):
        names.add(name)

names = (sorted(names))
len(names)

names[:5]

reporter_df = pd.DataFrame()
reporter_df["이름"] = names

reporter_df.shape

reporter_df.to_csv(reporter_path, index=False, encoding="utf-8-sig")
